user auth fix file roots
