+++ src/backend/schemas/authSchemas.js
@@ -1,0 +1,16 @@
+import Joi from 'joi';
+
+export const registerSchema = Joi.object({
+  email: Joi.string().email().required(),
+  password: Joi.string().min(8).required(),
+  repeatPassword: Joi.string().min(8).required(),
+});
+
+export const loginSchema = Joi.object({
+  email: Joi.string().email().required(),
+  password: Joi.string().min(8).required(),
+});
+
+export const emailSchema = Joi.object({
+  email: Joi.string().email().required(),
+});
